Edgar_FinishedNavigationMechanic
I finished the navigation mechanic with a pop up canvas to choose the location from. Audio clips play as well and there is a screen fade when teleporting. A controller input issue remains, so use the space button to open and close the navigation menu. And avoid pointing the hand ray on the navigation menu buttons when pressing space to close it. Space is recognized as a point click, so it will fade out but never fade in again, since the menu gets closed.